refactor(plugin-ai): move RunJS facade to client-v2
